code/requirements.txt
absl-py==1.1.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
kiwisolver==1.4.2
markdown==3.3.7
matplotlib==3.5.2
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyyaml==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
# torch==1.11.0+cu113
# torchaudio==0.11.0+cu113
# torchvision==0.12.0+cu113
tqdm==4.64.0
urllib3==1.26.9